+++ application/user_router.go
@@ -8,19 +8,20 @@ import (
 	controller "github.com/Lukmanern/gost/controller/user"
 	service "github.com/Lukmanern/gost/service/user"
 
-	rbacService "github.com/Lukmanern/gost/service/rbac"
+	permSvc "github.com/Lukmanern/gost/service/permission"
+	roleSvc "github.com/Lukmanern/gost/service/role"
 )
 
 var (
-	userPermService rbacService.PermissionService
-	userRoleService rbacService.RoleService
+	userPermService permSvc.PermissionService
+	userRoleService roleSvc.RoleService
 	userService     service.UserService
 	userController  controller.UserController
 )
 
 func getUserRoutes(router fiber.Router) {
-	userPermService = rbacService.NewPermissionService()
-	userRoleService = rbacService.NewRoleService(userPermService)
+	userPermService = permSvc.NewPermissionService()
+	userRoleService = roleSvc.NewRoleService(userPermService)
 	userService = service.NewUserService(userRoleService)
 	userController = controller.NewUserController(userService)
 	jwtHandler := middleware.NewJWTHandler()
